adjust all tests for suites tasks
Integration and unit tests for kpet have a lot of assets. Currently, to take
advantage of the pending kpet-db MR that gets rid of all task XML
snippets, 'maintainer: ' has to be added to the yamls. Field fetch_suffix
is not mandatory, but desirable.

The jinja template is adjusted to generate XML snippets instead of
loading them from files.

The rest is handled by the existing schema conversion.

V2:
* overhaul completely
V3:
* pylint

Signed-off-by: Jakub Racek <jracek@redhat.com>